requirements.txt
requests

pydantic
pydantic-settings
pydantic[email]

fastapi
uvicorn
gunicorn
starlette

sqladmin
sqlmodel
SQLAlchemy
sqlalchemy-utils
alembic
psycopg2-binary

httpx
Faker
pytest
coverage
pytest-cov
pytest-mock

bcrypt
passlib
python-dotenv
python-jose
itsdangerous
